hashing of integers for fingerprinting of patterns.
getVertexAndEdgeHistogramsP writes the fingerprint of a graph (being the bitwise
or of all hashed frequent edges in the graph) to pruning.

scanDB updates the fingerprint of a graph in subsequent levels and prunes
graphs that have been found to contain no frequent patterns from the level below.

still to do: candidate generation has to store fingerprint of frequent
patterns somehow