loading.c now also works for pipe input.
that is, writeCurrentGraph() now buffers the string representation of
the current graph in HEAD_PTR, VERTEX_PTR, and EDGE_PTR and
writeCurrentGraph() can then access these strings to output the string
representation of the current graph.
Before, this output was generated seeking around in the input stream,
which did work for files, but not for stdin, as this stream seems not
to be seekable.