addStringToSearchTreeRec now ensures that the edges in the resulting searchtree are stringMasters.
as search trees tend to live longer than the graphs they are derived from, this saves trouble, although it makes the addition of a string to the search tree depend on the length of the labels in each position of the string.

This has implications for all methods using addStringToSearchTreeRec, for example
addStringToSearchTree
addToSearchTree
etc.