implementation of a DFS graph mining algorithm.
at the moment it takes twice the time of a basic BFS graph mining (
also implemented here). mainBFS and mainDFS implement the two
strategies.

Testing is needed to see, if both methods return the same sets of
supports for same parameters.

At least, it seems to be possible to mine frequent trees in the set of
trees in the aids dataset, which is better than what was possible
before.