minHashing: speedup and sucky embedding operator inclusion
- 'ray of light' implements the anti-monotone pruning idea in the other
direction. i.e. if a pattern is found to be a match, all subgraphs of
it must match as well. so far, we only concluded for a non-match, that
all supergraphs are non-matches as well. this is currently only implemented
in fastMinHashForTrees

- hence, the EvaluationPlan now contains a reverse graph and we need a
way to compute it. reverseGraph() should probably be moved to graph.c

- fastMinHashForAndOr implements minHash for the shitty ANDOR amplification
embedding operator