make the local tree sampling algorithm work with disconnected graphs?
remove a speedup thing that does not process the root in its own rooted
components. this would reduce the number of characteristic computations
for roots v of v-rooted components but adds the burden of checking
whether the current root is the root of a connected component.

TODO might be worthwile to remove the extra handling of the 0-root now